feat: add suggest product in detail page
